+++ src/main/java/com/example/tily/_core/errors/CustomException.java
@@ -1,4 +1,4 @@
-package com.example.tily._core.errors.exception;
+package com.example.tily._core.errors;
 
 import com.example.tily._core.utils.ApiUtils;
 import lombok.Getter;
