+++ src/main/java/com/example/tily/_core/errors/GlobalValidationHandler.java
@@ -1,7 +1,5 @@
 package com.example.tily._core.errors;
 
-import com.example.tily._core.errors.exception.CustomException;
-import com.example.tily._core.errors.exception.ExceptionCode;
 import org.aspectj.lang.JoinPoint;
 import org.aspectj.lang.annotation.Aspect;
 import org.aspectj.lang.annotation.Before;
