+++ src/main/java/com/example/tily/comment/CommentService.java
@@ -1,7 +1,7 @@
 package com.example.tily.comment;
 
-import com.example.tily._core.errors.exception.ExceptionCode;
-import com.example.tily._core.errors.exception.CustomException;
+import com.example.tily._core.errors.ExceptionCode;
+import com.example.tily._core.errors.CustomException;
 import com.example.tily.alarm.Alarm;
 import com.example.tily.alarm.AlarmRepository;
 import com.example.tily.roadmap.Roadmap;
