+++ src/main/java/com/example/tily/image/ImageService.java
@@ -2,8 +2,8 @@ package com.example.tily.image;
 
 import com.example.tily._core.S3.FileFolder;
 import com.example.tily._core.S3.S3Service;
-import com.example.tily._core.errors.exception.CustomException;
-import com.example.tily._core.errors.exception.ExceptionCode;
+import com.example.tily._core.errors.CustomException;
+import com.example.tily._core.errors.ExceptionCode;
 import com.example.tily.roadmap.Roadmap;
 import com.example.tily.roadmap.RoadmapRepository;
 import com.example.tily.user.User;
